requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.115.5
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.48.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
Google-Images-Search==1.4.7
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.55.0
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.1
pydantic_core==2.27.1
pyfiglet==0.8.post1
pyparsing==3.2.0
python-dotenv==1.0.1
python-resize-image==1.1.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
termcolor==1.1.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.32.1
